Move build tools install
