+++ cli/src/commands/new/index.ts
@@ -17,13 +17,14 @@ import * as fs from "../../util/fs.js";
 import * as vi from "../../util/versioninfo.js";
 import { execFile } from "../../util/cp.js";
 import { isOnline } from "../../util/index.js";
-import { GitHubOwner, GitHubRepo, MinGoVersion, MinNodeVersion, MinTinyGoVersion, ModusHomeDir, SDK, parseSDK } from "../../custom/globals.js";
+import { MinGoVersion, MinNodeVersion, MinTinyGoVersion, SDK, parseSDK } from "../../custom/globals.js";
 import { withSpinner } from "../../util/index.js";
 import { extract } from "../../util/tar.js";
 import SDKInstallCommand from "../sdk/install/index.js";
 import { getHeader } from "../../custom/header.js";
 import * as inquirer from "@inquirer/prompts";
 import { getGoVersion, getTinyGoVersion } from "../../util/systemVersions.js";
+import { installBuildTools } from "../../util/installer.js";
 
 const MODUS_DEFAULT_TEMPLATE_NAME = "default";
 
@@ -250,7 +251,6 @@ export default class NewCommand extends Command {
 
     const sdkVersion = installedSdkVersion;
     const sdkPath = vi.getSdkPath(sdk, sdkVersion);
-
     const templatesArchive = path.join(sdkPath, "templates.tar.gz");
     if (!(await fs.exists(templatesArchive))) {
       this.logError(`Could not find any templates for ${sdkText} ${sdkVersion}`);
@@ -258,28 +258,7 @@ export default class NewCommand extends Command {
     }
 
     // Install build tools if needed
-    if (sdk == SDK.Go) {
-      const ext = os.platform() === "win32" ? ".exe" : "";
-      const buildTool = path.join(sdkPath, "modus-go-build" + ext);
-      if (!(await fs.exists(buildTool))) {
-        if (await isOnline()) {
-          const module = `github.com/${GitHubOwner}/${GitHubRepo}/sdk/go/tools/modus-go-build@${sdkVersion}`;
-          await withSpinner("Downloading the Modus Go build tool.", async () => {
-            await execFile("go", ["install", module], {
-              cwd: ModusHomeDir,
-              shell: true,
-              env: {
-                ...process.env,
-                GOBIN: sdkPath,
-              },
-            });
-          });
-        } else {
-          this.logError("Could not find the Modus Go build tool. Please try again when you are online.");
-          this.exit(1);
-        }
-      }
-    }
+    await installBuildTools(sdk, sdkVersion);
 
     // Create the app
     this.log(chalk.dim(`Using ${sdkText} ${sdkVersion}`));
